refactor(passenger-service): migrate IDs to BigInt and enhance validation/destructuring logic

+++ backend/services/passenger-service/prisma/migrations/20251102120710_init_bigint/migration.sql
@@ -3,12 +3,12 @@ CREATE TYPE "ClassType" AS ENUM ('economy', 'business', 'first');
 
 -- CreateTable
 CREATE TABLE "Passenger" (
-    "id" SERIAL NOT NULL,
+    "id" BIGSERIAL NOT NULL,
     "firstName" TEXT NOT NULL,
     "lastName" TEXT NOT NULL,
     "passportNumber" TEXT NOT NULL,
     "ticketNumber" TEXT NOT NULL,
-    "flightId" INTEGER NOT NULL,
+    "flightId" BIGINT NOT NULL,
     "seatNumber" TEXT,
     "classType" "ClassType" NOT NULL,
     "nationality" TEXT NOT NULL,
